Remove `spec` code from kazaam.go
